changes to api admin routes
