+++ server/api/items.js
@@ -1,5 +1,6 @@
 const router = require('express').Router()
 const Item = require('../db/models/item')
+const isAdmin = require('./admin')
 
 // http://expressjs.com/en/5x/api.html#router.param
 router.param('id', async (req, res, next, id) => {
@@ -44,6 +45,50 @@ router.put('/:id', async (req, res, next) => {
     let item = req.item
     await item.update(req.body)
     res.json(req.item)
+  } catch (error) {
+    next(error)
+  }
+})
+
+router.put('/:id', isAdmin, async (req, res, next) => {
+  try {
+    const item = await Item.findByPk(req.params.id)
+    if (item) {
+      await item.update(req.body)
+      res.json(item)
+    } else {
+      res.sendStatus(404)
+    }
+  } catch (error) {
+    next(error)
+  }
+})
+
+router.post('/', isAdmin, async (req, res, next) => {
+  try {
+    const {name, price, description, imageUrl, quantity} = req.body
+
+    let newItem = {name, price, admin: false}
+    if (description) newItem.description = description
+    if (imageUrl) newItem.imageUrl = imageUrl
+    if (quantity) newItem.quantity = quantity
+
+    const item = await Item.create(newItem)
+    res.json(item)
+  } catch (error) {
+    next(error)
+  }
+})
+
+router.delete('/:id', isAdmin, async (req, res, next) => {
+  try {
+    const item = await Item.findByPk(req.params.id)
+    if (item) {
+      res.send(item)
+      await item.destroy()
+    } else {
+      res.sendStatus(404)
+    }
   } catch (error) {
     next(error)
   }
