+++ server/api/users.js
@@ -1,5 +1,6 @@
 const router = require('express').Router()
 const {User, Order} = require('../db/models')
+const isAdmin = require('./admin')
 module.exports = router
 
 router.get('/me', (req, res, next) => {
@@ -9,5 +10,16 @@ router.get('/me', (req, res, next) => {
     else res.json(user)
   } catch (error) {
     next(error)
+  }
+})
+
+router.get('/allusers', isAdmin, async (req, res, next) => {
+  try {
+    const users = await User.findAll({
+      attributes: ['id', 'email', 'firstName', 'lastName', 'fullName']
+    })
+    res.json(users)
+  } catch (err) {
+    next(err)
   }
 })
